requirements.txt

ultralytics>=8.0.0
torch
torchvision
opencv-python
matplotlib
python-docx
